Use 64-bit as default
    Since the popcount is set up to 64-bit this prevents potential
    typecasting between 32 and 64 bit numbers.

+++ configuration_interaction/ci_helper.py
@@ -1,7 +1,7 @@
 import numba
 import numpy as np
 
-BITTYPE = np.uint32
+BITTYPE = np.uint64
 BITSTRING_SIZE = np.dtype(BITTYPE).itemsize * 8
 
 # Const used by the Hamming weight algorithm
